requirements.txt
bitarray==3.4.0
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.58.0
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pillow==11.2.1
plum-dispatch==1.7.4
pycryptodome==3.22.0
pyparsing==3.2.3
pypdf==5.5.0
python-dateutil==2.9.0.post0
regex==2024.11.6
six==1.17.0
spire-doc==13.4.6
tksheet==7.5.12
tqdm==4.67.1
typing_extensions==4.13.2
